+++ ninja-test-utilities/src/main/java/ninja/NinjaFluentLeniumTest.java
@@ -20,7 +20,6 @@ import org.fluentlenium.adapter.junit.FluentTest;
 import org.junit.After;
 import org.junit.Before;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.htmlunit.HtmlUnitDriver;
 
 import com.google.inject.Injector;
 
@@ -28,61 +27,66 @@ import ninja.utils.NinjaTestServer;
 
 public abstract class NinjaFluentLeniumTest extends FluentTest {
 
-	public WebDriver webDriver = new HtmlUnitDriver();
-	
-	public NinjaTestServer ninjaTestServer;
-	
+    /**
+     * @deprecated The driver should be access via {@link #getDriver()}. Null
+     *             variable left for deprecation notice only.
+     * @see #getDriver()
+     */
+    @Deprecated
+    public WebDriver webDriver;
+
+    public NinjaTestServer ninjaTestServer;
+
     @Before
     public void startupServer() {
         ninjaTestServer = new NinjaTestServer();
     }
-	
-	  /**
-	   * @deprecated Updating to FluentLenium 3.2.0, the driver to use is now either a string,
-	   * returned by overriding the {@link #getWebDriver()} method, or the driver implementation,
-	   * by overriding the {@link #newWebDriver()} method.
-	   * @see #getWebDriver()
-	   * @see #newWebDriver()
-	   */
-	  @Deprecated
+
+    /**
+     * @deprecated Updating to FluentLenium 3.2.0, the driver to use is now
+     *             either a string, returned by overriding the
+     *             {@link #getWebDriver()} method, or the driver implementation,
+     *             by overriding the {@link #newWebDriver()} method.
+     * @see #getWebDriver()
+     * @see #newWebDriver()
+     */
+    @Deprecated
     public WebDriver getDefaultDriver() {
         return webDriver;
     }
-    
-    /**
-     * @see org.fluentlenium.adapter.FluentAdapter#newWebDriver()
-     */
+
     @Override
-    public WebDriver newWebDriver() {
-        return getDefaultDriver();
+    public String getWebDriver() {
+        return "htmlunit";
     }
-    
+
     /**
      * @see #getBaseUrl()
      */
     @Deprecated
     public String getServerAddress() {
-    	return ninjaTestServer.getServerAddress();
+        return ninjaTestServer.getServerAddress();
     }
 
     /**
      * Will be automatically used within the goTo() method.
+     * 
      * @see org.fluentlenium.adapter.FluentAdapter#getBaseUrl()
      */
     @Override
     public String getBaseUrl() {
         return ninjaTestServer.getBaseUrl();
     }
-    
+
     public String getServerUrl() {
         return ninjaTestServer.getServerUrl();
     }
-    
+
     @After
     public void shutdownServer() {
         ninjaTestServer.shutdown();
     }
-    
+
     public Injector getInjector() {
         return ninjaTestServer.getInjector();
     }