+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/src/main/java/controllers/ApplicationController.java
@@ -44,7 +44,6 @@ import ninja.Results;
 
 import com.google.common.collect.Maps;
 import com.google.inject.Inject;
-import com.google.inject.Singleton;
 
 import dao.ArticleDao;
 import dao.SetupDao;