+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/src/main/java/controllers/ApiController.java
@@ -45,7 +45,6 @@ import ninja.SecureFilter;
 import ninja.params.PathParam;
 
 import com.google.inject.Inject;
-import com.google.inject.Singleton;
 
 import dao.ArticleDao;
 import etc.LoggedInUser;