+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/src/main/java/models/Article.java
@@ -30,7 +30,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.collect.Lists;
 
 @Entity