+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,4 +1,5 @@
 * 2018-09-09 JPA Blog archetype cleaned up (jlannoy)
+* 2018-09-09 Jetty context configuration fixed (jlannoy)
 * 2018-08-23 Updated libraries (jlannoy)
     * com.fasterxml.jackson.core:jackson-* ...................... 2.8.1 -> 2.9.6
     * com.google.guava:guava .................................... 19.0 -> 26.0-jre